Safari-609.1.11.1/LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE MESSAGE: line 14: TypeError: Type error
PASS: Invalid blend modes do not trigger a crash. 
